[Move(lotto)] 패키지 이동
